Stop systemd service if gitlab-runner is not healthy
When gitlab-runner is deemed unhealthy the process does not terminate on
its own. Instead it just prints an error:

    ERROR: Runner https://gitlab.com/zcw… is not healthy and will be disabled!

As far as I can tell this failure is not reported through Prometheus
/metrics endpoint and can be detected only by executing `gitlab-runner verify`.
That's why we add ExecStartPost= to gitlab-runner.service. Verification
script gets started immediately after invoking ExecStart= (for Type=simple units)
and until this script finishes successfully the service is not considered
active by systemd: try-restart will not trigger a restart while the
service is in activating state.

By default gitlab-runner.service includes a Restart=always condition, but
this is not desirable on our short-lived runner instances. Instead of
restarting potentially unhealthy service we replace the whole instance on
which gitlab-runner is executed.