Do not reserve a static external IPv4
This does not affect pricing much but at least it makes Terraform
dependency graph a little simpler.

Static and dynamic IPv4 addresses at Yandex Cloud are priced differently
only when they are not attached to any VM (static address incurs extra
reservation costs while dynamic one just goes away automatically).
When an address is attached to a VM the price is the same regardless of
whether that address is static or a dynamic.

https://cloud.yandex.com/en-ru/docs/vpc/pricing#prices-public-ip